% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatistics.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Compute Statistics On iChart}
\usage{
computeStatistics(iChart, startWindow = 300, endWindow = 3000,
  save_results = F)
}
\arguments{
\item{iChart}{A data frame in iChart format with iChart column names.}

\item{startWindow}{An integer indicating the lower bound of the analysis window in milliseconds.}

\item{endWindow}{An integer indicating the upper bound of the analysis window in milliseconds.}

\item{save_results}{A boolean indicating whether the results should be saved to disk.}
}
\description{
\code{computeStatistics()} takes two values that indicated an analysis window and computes
  the following statistics for each trial: RT, longestLook, longestGap, Gap, longestGapPosition,
  NumberOfShifts, FirstFixation.
}
\details{
This function computes the relevant stats on an iChart.
}
\examples{
\dontrun{d <- computeStatistics(iChart, startWindow=0, endWindow=3000, save_results = TRUE)}
}
