% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlots.R
\name{createPlots}
\alias{createPlots}
\title{Create LWL Plots}
\usage{
createPlots(iChart, startWindow, endWindow, RejectLongestGap, RejectFirstGap,
  RejectRT, plotStats = "PP", group = "", ...)
}
\arguments{
\item{iChart}{A data frame in iChart format with iChart column names.}

\item{startWindow}{An integer indicating the lower bound of the analysis window in milliseconds.}

\item{endWindow}{An integer indicating the upper bound of the analysis window in milliseconds.}

\item{RejectLongestGap}{A boolean indicating whether bad longest gaps should be filtered out of the computation.}

\item{RejectFirstGap}{A boolean indicating whether bad first gaps should be filtered out of the computation.}

\item{RejectRT}{A boolean indicating whether bad RTs should be filtered out of the computation.}

\item{plotStats}{A string indicating which type of plot to generate measure to use (PP, OC, OC_T, OC_D).}

\item{group}{A string indicating the variable name for any grouping variables (e.g., conditions).}

\item{...}{Additional style arguments for tweaking the ggplot output.}
}
\description{
\code{createPlots()} generates either a Profile Plot or Onset Contingency Plot.
}
\details{
This function generates two types of plots for PEEK data: Profile Plots and Onset Contingency Plots
}
\examples{
\dontrun{acc <- createPlots(d, startWindow=0, endWindow=3000,
  RejectFirstGap=FALSE, RejectLongestGap=FALSE, RejectRT=FALSE,
  plotStats = "PP", group = "")}

}
