% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineOnset.R
\name{defineOnset}
\alias{defineOnset}
\title{Define Onsets For iChart}
\usage{
defineOnset(iChart, critonset = 0, includeAways = FALSE)
}
\arguments{
\item{iChart}{A data frame in iChart format with iChart column names.}

\item{critonset}{An integer indicating the critical onset value in milliseconds.}

\item{includeAways}{A boolean indicating whether Away responses should be inlcluded.}
}
\description{
\code{defineOnset()}the type of trial (D-, T-, or A-initial) based on a user-specified
  critical onset value provided in milliseconds.
}
\details{
This function defines the type of trial (D-, T-, or A-initial) based on a user-specified
critical onset value.
}
\examples{
\dontrun{d <- defineOnset(iChart, critonset = 0, includeAways = FALSE)}
}
