% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeiChart.R
\name{describeiChart}
\alias{describeiChart}
\title{Describe iChart}
\usage{
describeiChart(iChart)
}
\arguments{
\item{iChart}{A data frame created by the \code{readiChart()} function.}
}
\description{
\code{describeiChart()} provides a quick check of the number of trials
  for each participant and condition. It is also useful for checking the effects of any data
  filtering.
}
\details{
This function returns a tabular summary of trial and prescreening counts for each participant and condition.
}
\examples{
\dontrun{d <- describeiChart(iChart)}
}
