% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteriChart.R
\name{filteriChart}
\alias{filteriChart}
\title{Filter iChart}
\usage{
filteriChart(iChart, minRT, maxRT, maxfirstgap, maxlonggap,
  save_results = FALSE)
}
\arguments{
\item{iChart}{A data frame in iChart format with iChart column names.}

\item{minRT}{An integer indicating the lower bound of the analysis window in milliseconds.}

\item{maxRT}{An integer indicating the upper bound of the analysis window in milliseconds.}

\item{maxfirstgap}{An integer indicating the upper bound of the first gap shift length frames.}

\item{maxlonggap}{An integer indicating the upper bound of the longest gap in a trial in frames.}

\item{save_results}{A boolean indicating whether the results should be saved to disk.}
}
\description{
\code{filteriChart()} takes two values that indicated an analysis window and computes
  the following statistics for each trial: RT, longestLook, longestGap, Gap, longestGapPosition,
  NumberOfShifts, FirstFixation.
}
\details{
This function creates variables that can be used to
filter trials out of the iChart based on user-specified exlcusionary criteria.
}
\examples{
\dontrun{d <- filteriChart(iChart, minRT=300, maxRT=1800, maxfirstgap=15,
  maxlonggap=15, save_results = TRUE)}

}
