% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPrescreened.R
\name{filterPrescreened}
\alias{filterPrescreened}
\title{Filter Prescreened Trials From iChart}
\usage{
filterPrescreened(iChart, save_results = TRUE)
}
\arguments{
\item{iChart}{A data frame created by the \code{readiChart()} function.}

\item{save_results}{A boolean indicating whether the results should be saved to disk.}
}
\description{
\code{describeiChart()} provides a quick check of the number of trials
  for each participant and condition. It is also useful for checking the effects of any data
  filtering.
}
\details{
This function returns an iChart with prescreened out trials removed
}
\examples{
\dontrun{d <- filterPrescreened(iChart, save_results = TRUE)}
}
