% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphValues.R
\name{generateGraphValues}
\alias{generateGraphValues}
\title{Generate Graph Values}
\usage{
generateGraphValues(iChart, filter_criteria = list("GoodRT", "GoodFirstGap",
  "GoodLongestGap"), group_cols = list("Sub.Num", "Condition"),
  startWindow = 0, endWindow = 3000, save_results = TRUE)
}
\arguments{
\item{iChart}{A data frame in iChart format with iChart column names.}

\item{filter_criteria}{A list of strings indicating the variables that should be used for filtering iChart. Options include:
"GoodRT", "GoodFirstGap", and "GoodLongestGap." Or if you don't want to filter pass the string "none"}

\item{group_cols}{A list of strings indicating the variable names  for any grouping variables.}

\item{startWindow}{An integer indicating the lower bound of the analysis window in milliseconds.}

\item{endWindow}{An integer indicating the upper bound of the analysis window in milliseconds.}

\item{save_results}{A boolean indicating whether the results should be saved to disk.}
}
\description{
\code{generateGraphValues()} computes the mean proportion looking each participant, condition, and frame.
}
\details{
This function aggregates iChart Proportion Looking for each trial and Accuracy for each participant, condition,
and frame, and returns this information in a data frame.
}
\examples{
\dontrun{df_gvs <- generateGraphValues(d_analysis,
    filter_criteria = list("GoodRT", "GoodFirstGap", "GoodLongestGap"),
    group_cols = list("Sub.Num", "Condition"),
    startWindow = 0, endWindow = 3000,
    save_results = FALSE)}

}
