% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePPgraphValues.R
\name{generatePPgraphValues}
\alias{generatePPgraphValues}
\title{Generate Profile Plot Graph Values}
\usage{
generatePPgraphValues(iChart, filter_criteria = list("GoodFirstGap",
  "GoodLongestGap"), group_cols = list("Condition"), gvStartWindow = 0,
  gvEndWindow = 3000, save_results = TRUE)
}
\arguments{
\item{iChart}{A data frame in iChart format with iChart column names.}

\item{filter_criteria}{A list of strings indicating the variables that should be used for filtering iChart. Options include:
"GoodRT", "GoodFirstGap", and "GoodLongestGap." Or if you don't want to filter pass the string "none"}

\item{group_cols}{A list of strings indicating the variable names  for any grouping variables.}

\item{gvStartWindow}{An integer indicating the lower bound of the graph values window in milliseconds.}

\item{gvEndWindow}{An integer indicating the upper bound of the graph values window in milliseconds.}

\item{save_results}{A boolean indicating whether the results should be saved to disk.}
}
\description{
\code{generatePPgraphValues()} computes the mean proportion looking each participant, condition, and frame.
}
\details{
This function aggregates iChart Proportion Looking for each trial and Accuracy for each participant, condition,
and frame, and returns this information in a data frame.
}
\examples{
\dontrun{df_gvs <- generatePPgraphValues(d_analysis,
    filter_criteria = list("GoodRT", "GoodFirstGap", "GoodLongestGap"),
    group_cols = list("Condition"),
    gvStartWindow = 0, gvEndWindow = 3000,
    save_results = FALSE)}

}
