% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiChart.R
\name{readiChart}
\alias{readiChart}
\title{Read iChart}
\usage{
readiChart(iChartFile, sampling_rate = 17)
}
\arguments{
\item{iChartFile}{Either a path to a file, a connection, or literal data.}

\item{sampling_rate}{An integer indicating the sampling rate of the eyetracker (17 vs. 30).}
}
\description{
\code{readiChart()} is a special case of \code{read.table()}. It's useful for doing minor data
   processing on the raw .text file that is specific to the Language Learning Lab's analysis workflow.
}
\details{
This function allows you to read a wide format data file generated by RTobii as an iChart
}
\examples{
\dontrun{d <- readiChart(iChartFile = "Habla2_25_iChart_wide.txt", sampling_rate = 17)}
}
