% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeUnknownWords.R
\name{removeUnknownWords}
\alias{removeUnknownWords}
\title{Remove Unknown Words From iChart}
\usage{
removeUnknownWords(iChart, df_unknown_words)
}
\arguments{
\item{iChart}{A data frame created by the \code{readiChart()} function.}

\item{df_unknown_words}{A data frame of unknown words for each participant that should be filtered from the analysis.}
}
\description{
\code{removeUnknownWords()} provides a quick way to rename conditions in an iChart
}
\details{
This function removes unknown words for each participant
}
\examples{
\dontrun{d <- removeUnknownWords(iChart, df_unknown_words)}

}
