% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameCondition.R
\name{renameCondition}
\alias{renameCondition}
\title{Rename Condition In iChart}
\usage{
renameCondition(iChart, oldCondition, newCondition)
}
\arguments{
\item{iChart}{A data frame created by the \code{readiChart()} function.}

\item{oldCondition}{A string indicating the old condition name}

\item{newCondition}{A string indicating the new condition name}
}
\description{
\code{renameCondition()} provides a quick way to rename conditions in an iChart
}
\details{
This function renames the conditions in an iChart
}
\examples{
\dontrun{d <- renameCondition(iChart, oldCondition = "Experimental", newCondition = "Vanilla")}
}
