% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeProfilePlot.R
\name{makeProfilePlot}
\alias{makeProfilePlot}
\title{Make Profile Plot}
\usage{
makeProfilePlot(df_gvs, startWindow, endWindow, bin_width = 17)
}
\arguments{
\item{df_gvs}{A data frame of graph values created by the \code{generateGraphValues()} function.}

\item{startWindow}{An integer indicating the lower bound of the plotting window in milliseconds.}

\item{endWindow}{An integer indicating the upper bound of the plotting window in milliseconds.}

\item{bin_width}{An integer indicating the sampling rate or smoothing rate for the plot. Values
currently supported include 17 or 33 ms.}
}
\description{
\code{makeProfilePlot()} creates a profile plot
}
\details{
This function takes graph values and generates a profile plot (proportion looking to target as a function of
time).
}
\examples{
\dontrun{makeProfilePlot(df_gvs, startWindow = 0, endWindow = 3000, bin_width = 17)}

}
