% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeUnknownWords.R
\name{removeUnknownWords}
\alias{removeUnknownWords}
\title{Remove Unknown Words From iChart}
\usage{
removeUnknownWords(iChart, knownWords_file, knows_threshold = 3)
}
\arguments{
\item{iChart}{A data frame created by the \code{readiChart()} function.}

\item{knownWords_file}{A file with a table of unknown words for each participant}

\item{knows_threshold}{An integer indicating the threshold value for whether a word is known.}
}
\description{
\code{removeUnknownWords()} provides a quick way to rename conditions in an iChart
}
\details{
This function removes unknown words for each participant
}
